XDMA: libxdma: fix setting adjacent descriptors
Use the next desc's address to do alignment and not the assumption that
the first descriptor will be aligned.

Introduce xdma_get_next_adj to calculate the number of adjacent
descriptors based on the remaining number of descriptors, the maximum
number of adjacent descriptors and the remaining number of descriptors
before the end of the 4k page.